+++ configuration/src/main/java/org/ehrbase/configuration/config/security/SecurityConfigNoOp.java
@@ -22,7 +22,6 @@ import javax.annotation.PostConstruct;
 import org.ehrbase.service.IAuthenticationFacade;
 import org.springframework.boot.actuate.autoconfigure.endpoint.web.WebEndpointProperties;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.autoconfigure.security.SecurityProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Primary;
 import org.springframework.security.authentication.AnonymousAuthenticationToken;
@@ -75,8 +74,9 @@ public final class SecurityConfigNoOp extends SecurityConfig {
      * the app with security enabled to be able to use an external oauth2 client.
      */
     @Bean
-    public InMemoryUserDetailsManager inMemoryUserDetailsManager(SecurityProperties properties) {
-        final SecurityProperties.User user = properties.getUser();
+    public InMemoryUserDetailsManager inMemoryUserDetailsManager(
+            org.springframework.boot.autoconfigure.security.SecurityProperties properties) {
+        final org.springframework.boot.autoconfigure.security.SecurityProperties.User user = properties.getUser();
         final List<String> roles = user.getRoles();
         return new InMemoryUserDetailsManager(User.withUsername(user.getName())
                 .password(user.getPassword())