+++ configuration/src/test/java/org/ehrbase/configuration/cors/CorsIT.java
@@ -18,30 +18,36 @@
 package org.ehrbase.configuration.cors;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.options;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import org.junit.jupiter.api.Disabled;
+import org.ehrbase.configuration.test.EhrbaseConfigurationIntegrationTest;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.web.servlet.MockMvc;
 
-@SpringBootTest(properties = {"spring.cache.type=simple"})
-@AutoConfigureMockMvc
-@Disabled
-class CorsNoAuthIT {
+class CorsIT {
 
-    @Autowired
-    private MockMvc mockMvc;
+    @EhrbaseConfigurationIntegrationTest
+    abstract class CursBaseTest {
 
-    @Test
-    void testCors() throws Exception {
-        mockMvc.perform(options("/rest/openehr/v1/definition/template/adl1.4")
-                        .header("Access-Control-Request-Method", "GET")
-                        .header("Origin", "https://client.ehrbase.org"))
-                .andDo(print())
-                .andExpect(status().isOk());
+        @Autowired
+        private MockMvc mockMvc;
+
+        @Test
+        void OPTION_corsHeaderPresent_OK() throws Exception {
+            mockMvc.perform(options("/rest/openehr/v1/definition/template/adl1.4")
+                            .header("Access-Control-Request-Method", "GET")
+                            .header("Origin", "https://client.ehrbase.org"))
+                    .andExpect(status().isOk());
+        }
     }
+
+    @Nested
+    class NoAuth extends CursBaseTest {}
+
+    @TestPropertySource(properties = {"security.authType=basic"})
+    @Nested
+    class BasicAuth extends CursBaseTest {}
 }