+++ CHANGELOG.md
@@ -7,9 +7,26 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
  ### Added
-* Added `STORED_QUERY_CACHE` ([#1258](https://github.com/ehrbase/ehrbase/pull/1258))
 * Support for AQL on `VERSION.commit_audit.committer` paths ([#1264](https://github.com/ehrbase/ehrbase/pull/1264))
  ### Changed 
+ ### Fixed 
+
+## [2.2.0]
+### Added
+* Added AQL debug support ([#1296](https://github.com/ehrbase/ehrbase/pull/1296))
+### Changed 
+ - Upgrade openEHR_SDK to version 2.11.0 see https://github.com/ehrbase/openEHR_SDK/blob/develop/CHANGELOG.md
+* The field `q` of AQL query responses now contain the requested, and not the executed, query string  ([#1296](https://github.com/ehrbase/ehrbase/pull/1296))
+* The field `meta._schema_version` of AQL query responses has been changed to `1.0.3`  ([#1296](https://github.com/ehrbase/ehrbase/pull/1296))
+* Return HTTP 422 Unprocessable Content in case fetch or offset is defined inside the aql query and as parameter ([#1325](https://github.com/ehrbase/ehrbase/pull/1325)).
+### Fixed
+
+## [2.1.0]
+ ### Added
+* Added `STORED_QUERY_CACHE` ([#1258](https://github.com/ehrbase/ehrbase/pull/1258))
+* Added new config option `ehrbase.security.management.endpoints.web.csrf-validation-enabled` ([#1294](https://github.com/ehrbase/ehrbase/pull/1294),[#1297](https://github.com/ehrbase/ehrbase/pull/1297))
+ ### Changed 
+ - Upgrade openEHR_SDK to version 2.10.0 see https://github.com/ehrbase/openEHR_SDK/blob/develop/CHANGELOG.md 
 * Changed `StoredQueryRepository` methods to only accept `StoredQueryQualifiedName` as arguments ([#1258](https://github.com/ehrbase/ehrbase/pull/1258))
  ### Fixed 
 * Fixed an issue with AQL, which caused NPEs when the query required adding filtering subqueries on a DV_ORDERED path ([#1293](https://github.com/ehrbase/ehrbase/pull/1293))
@@ -21,4 +38,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   
   See [UPDATING.md](./UPDATING.md) for details on how to update to the new release.
 
-[unreleased]: https://github.com/ehrbase/ehrbase/compare/v2.0.0...HEAD
+[2.1.0]: https://github.com/ehrbase/ehrbase/compare/v2.0.0...v2.1.0
+[2.2.0]: https://github.com/ehrbase/ehrbase/compare/v2.1.0...v2.2.0
+[unreleased]: https://github.com/ehrbase/ehrbase/compare/v2.2.0...HEAD